Fix indentation that is breaking the pluging execution
The first line of the DOCUMENTATION section in the plugis in not
indented properly. This is resulting in the ansible lookup attempts
failing. Had to fix the indentation to gt the ansible-playbooks to
work properly. Attached below is a gist of the error thrown by ansible.
https://gist.github.com/sushilsuresh/586c99db5cf1968931f46b419d3ae137

The version of key software used
ansible: 2.9.0

pip mocule used.
secret-server-sdk-client==1.0